#266 Refactor to use ala-ws-security library for auth handling
Replaces local TokenService, TokenClient, TokenInterceptor, and Pac4jContextProvider implementations with those from the ala-ws-security library. Updates dependency versions and imports accordingly, and removes now-unnecessary local auth utility classes. This change centralizes authentication logic and reduces code duplication.
