+++ lists-service/pom.xml
@@ -56,12 +56,12 @@
     <dependency>
       <groupId>au.org.ala</groupId>
       <artifactId>ala-ws-spring-security</artifactId>
-      <version>7.2.0-SNAPSHOT</version>
+      <version>7.3.0-SNAPSHOT</version>
     </dependency>
     <dependency>
       <groupId>au.org.ala</groupId>
       <artifactId>ala-ws-security</artifactId>
-      <version>7.2.0-SNAPSHOT</version>
+      <version>7.3.0-SNAPSHOT</version>
     </dependency>
     <dependency>
       <groupId>com.fasterxml.jackson.module</groupId>
