+++ lists-service/src/main/java/au/org/ala/listsapi/SecurityConfig.java
@@ -17,7 +17,7 @@ import org.springframework.security.web.firewall.HttpFirewall;
 
 @Configuration
 @EnableWebSecurity
-@ComponentScan(basePackages = {"au.org.ala.ws.security"})
+@ComponentScan(basePackages = {"au.org.ala.ws.security", "au.org.ala.security.common"})
 @EnableGlobalMethodSecurity(securedEnabled = true)
 @EnableCaching
 @Order(1)
