--- lists-service/src/main/java/au/org/ala/listsapi/util/auth/Pac4jContextProvider.java
@@ -1,13 +1,0 @@
-package au.org.ala.listsapi.util.auth;
-
-import org.pac4j.core.context.WebContext;
-
-// TODO: move this and related classes into ala-security-project
-
-/**
- * Provides a Pac4j Context via static methods or similar so that the client code need not take them as params.
- */
-public interface Pac4jContextProvider {
-
-    WebContext webContext();
-}
